channel_colors [nfc]: Rename files to channel_colors
Fixes parts of #631
